requirements.txt
dj-database-url==0.5.0
Django==2.1
django-cors-headers==2.4.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
ForgeryPy3==0.3.1
gunicorn==19.9.0
pkg-resources==0.0.0
psycopg2-binary==2.7.5
PyJWT==1.6.4
python-decouple==3.1
pytz==2018.5
whitenoise==4.0
